consolidated the password prompts into 1 file
