+++ internal/providers/wifi/README.md
@@ -12,10 +12,14 @@ One of the following utils/backend:
 
 #### Password Prompts
 
-When connecting to a new password-protected network, a password prompt is used to request the password:
+When connecting to a new password-protected network, a password prompt is used to request the password.
 
+If the configured prompt is not available, all presets are tried, with `terminal` as the final fallback.
+
+- `walker` - Walker in dmenu mode
+- `rofi` - Rofi in dmenu mode
+- `wofi` - Wofi in dmenu mode
 - `terminal` - Opens a terminal window for password input
-- `dmenu` - Uses a dmenu-compatible tool (walker, rofi, wofi)
 - `custom` - Uses a custom command defined in `custom_prompt_command`
 
 #### Configuration
@@ -23,14 +27,11 @@ When connecting to a new password-protected network, a password prompt is used t
 - `backend` string - WiFi backend (default: `auto`)
   - `auto` - Will try all backends until one works
   - `nm` - Network Manager
-- `password_prompt` string - Password prompt method (default: `dmenu`)
-  - `auto` - Will try all prompts until one works
-  - `terminal` - Terminal-based password prompt
-  - `dmenu` - dmenu-compatible tool
-  - `custom` - Custom command
-- `dmenu_command` string - dmenu-compatible tool to use (default: `walker`)
-  - `auto` - Will try walker, rofi, wofi in order
+- `password_prompt` string - Password prompt method (default: `walker`)
+  - `auto` - Will try all presets, then fall back to terminal
   - `walker`, `rofi`, `wofi` - Use a specific tool
+  - `terminal` - Terminal-based password prompt
+  - `custom` - Custom command
 - `custom_prompt_command` string - Custom command for the `custom` password prompt. Use `%PROMPT%` as placeholder for the prompt text
   - e.g. `rofi -dmenu -password -p %PROMPT%`
 - `message_time` int - Seconds to show status messages (default: `1`)
